+++ MCM_constants.py
@@ -1,412 +1,0 @@
-#Script containing all relevant MCM modules. This file was created from a manual copy of the
-#kpp_constants.f90 file found on mcm.leeds.ac.uk/MCM folder 23/3/17
-
-import numpy
-
-def mcm_constants(time, temp, H2O):
-    # calculates rate constants from arrhenius informtion
-    # Documentation with MCM does not indicate unites of time or temp!!
-
-    # F90 archive - REAL(dp) time, temp, M, N2, O2, RO2, H2O
-
-    # F90 archive - OPEN(13,file='photolysis.txt', status='old')
-
-    # calculate zenith angle for time of day
-    zenith_dict=zenith(time)
-    theta=zenith_dict['theta']
-    secx=zenith_dict['secx']
-    cosx=zenith_dict['cosx']
-
-    #return a dictionary containing all relevant variables
-    mcm_constants_dict=dict()
-
-    # ************************************************************************
-    # define generic reaction rates.
-    # ************************************************************************
-
-    # constants used in calculation of reaction rates
-    M  = 2.55E+19  #Check this against pressure assumed in model
-    N2 = 0.79*M
-    O2 = 0.2095*M
-
-    # kro2no : ro2      + no      = ro      + no2
-    #        : ro2      + no      = rono2
-    # iupac 1992
-    kro2no    = 2.40E-12*numpy.exp(360.0/temp)
-    mcm_constants_dict['KRONO2']=kro2no
-
-    # kro2ho2: ro2      + ho2     = rooh    + o2
-    # mcm protocol [1997]
-    kro2ho2   = 2.91E-13*numpy.exp(1300.0/temp)
-    mcm_constants_dict['KRO2HO2']=kro2ho2
-
-    # kapho2 : rcoo2    + ho2     = products
-    # mcm protocol [1997]
-    kapho2    = 4.30E-13*numpy.exp(1040.0/temp)
-    mcm_constants_dict['KAPHO2']=kapho2 
-
-    # kapno  : rcoo2    + no      = products
-    # mej [1998]
-    kapno = 8.10E-12*numpy.exp(270.0/temp)
-    mcm_constants_dict['KAPNO']=kapno 
-
-    # kro2no3: ro2      + no3     = products
-    # mcm protocol [1997]
-    kro2no3   = 2.50E-12
-    mcm_constants_dict['KRO2NO3']=kro2no3
-
-    # kno3al : no3      + rcho    = rcoo2   + hno3
-    # mcm protocol [1997]
-    kno3al    = 1.44E-12*numpy.exp(-1862.0/temp)
-    mcm_constants_dict['KNO3AL']=kno3al 
-
-    # kdec   : ro                 = products
-    # mcm protocol [1997]
-    kdec      = 1.00E+06
-    mcm_constants_dict['KDEC']=kdec
-
-    krosec = 1.50E-14*numpy.exp(-200.0/temp)
-    mcm_constants_dict['KROSEC']=krosec
-
[…]
