+++ docs/index.md
@@ -1,0 +1,184 @@
+# PyBox
+
+PyBox is a Python based box-model generator and simulator designed for atmospheric chemistry and aerosol studies. The first phase of the PyBox project is to develop a gas phase model, using the reaction information within the [Master Chemical Mechanism (MCM)](http://mcm.leeds.ac.uk/MCM/) as the basis, coupled with an idealised sectional aerosol model. PyBox also relates component properties, using molecular structural information, through the [UManSysProp](http://umansysprop.seaes.manchester.ac.uk) informatics suite.  Any public release will occur according to new processes added, agreement from any partner contributions and/or associated peer-review papers.
+
+Please check the project wiki page for more information on updates and planned releases. 
+
+This project is licensed under the terms of the GNU General Public License v3.0, as provided with this repository. 
+
+# Table of contents
+1. [Model overview](#Model-overview)
+2. [Dependencies](#Dependencies)
+3. [Folder structure and running the model](#Folder-Structure)
+4. [Unit Tests](#Automated-unit-tests)
+5. [Contributing](#Contributing)
+6. [Code of Conduct](#Code-of-Conduct)
+
+## Model overview<a name="Model-overview"></a>
+
+PyBox works on the basis of reading a file that defines reactions between compounds in the gas phase and the associated reaction coefficient. For example, take the MCM [Alpha-Pinene](https://en.wikipedia.org/wiki/Alpha-Pinene) chemical mechanism file  'MCM_APINENE.eqn.txt' stored in the 'mechanism_files' directory of PyBox. This contains the following snippet of text:
+
+     {125.} 	 C96OOH + OH = C96O2 : 	1.90D-12*EXP(190/TEMP) 	;
+     {126.} 	 C96OOH + OH = NORPINAL + OH : 	1.30D-11 	;
+     {127.} 	 C96OOH = C96O + OH : 	J(41)+J(22) 	;
+     {128.} 	 C96NO3 + OH = NORPINAL + NO2 : 	2.88D-12 	;
+     {129.} 	 C96NO3 = C96O + NO2 : 	J(53)+J(22) 	;
+     {130.} 	 C96O = C97O2 : 	4.20D+10*EXP(-3523/TEMP) 	;
+     {131.} 	 C96OH + OH = NORPINAL + HO2 : 	7.67D-12 	;
+     {132.} 	 C96OH = C96O + HO2 : 	J(22) 	;
+
+Where the equation number is defined first, then the reactants/products along with a defined rate coefficient. This equation file is parsed by functions in 'Parse_eqn_file.py', providing information that can be used to set up and solve the relevant ordinary differential equations (ODEs) to simulate the evolution of the chemical mechanism.  Each component in this chemical mechanism also has an associated record of chemical structure in the form of a [SMILES string](http://www.daylight.com/dayhtml/doc/theory/theory.smiles.html). This information is carried in a .xml file, provided by the MCM, and stored in the root directory of PyBox. Why is this important? Well, this information is taken by the [UManSysProp](http://umansysprop.seaes.manchester.ac.uk) informatics suite and allows us to predict properties of each compound that helps us predict whether they are likely to remain in the gas phase or condense to an existing particulate phase through gas-to-particle partitioning. Before we take a look at the directory structure provided in this repository, lets deal with the dependencies.
+
+## Dependencies<a name="Dependencies"></a>
+
+PyBox has been built in the [Anaconda Python 3.6 environment](https://www.anaconda.com/download/#macos). Utilised modules included within the Anaconda environment are:
+
+ - Numpy
+ - Scipy
+ - Matplotlib
+ - unittest
+ - shutil
+ - xml.dom
+ - multiprocessing
+ 
+Additional dependencies are:
+
[…]
