### 2018-05-23 #750 @loftytopping
Hi! Great. Ok cool, can you get in touch with 'dvalters' from the current JOSS reviewers list too? This covers Python, OpenMP and the domain of interest so its perfect for PyBox if they can review :).
