### 2018-05-25 #750 @lheagy
👋 Hi @dvalters, would you be willing to review this submission for JOSS?
