### 2018-05-29 #755 @whedon
[REVIEW]: PyBox: An automated box-model generator for atmospheric chemistry and aerosol simulations.
**Submitting author:** @loftytopping (<a href="http://orcid.org/0000-0001-8247-9649">David Topping</a>)
**Repository:** <a href="https://github.com/loftytopping/PyBox" target ="_blank">https://github.com/loftytopping/PyBox</a>
**Version:** v1.0.0
