### 2018-06-02 loftytopping/PyBox#3 @dvalters
[joss-review] Release version/tag needed
Cross ref from openjournals/joss-reviews#755:

To satisfy the checkbox for the:

>  Version: Does the release version given match the GitHub release (v1.0.0)?

I think there needs to be a release v1.0.0 on the github repo, either a tag or a formal release here. (At the minute v1.0.0 links to the master branch rather than a tag.)
