### 2018-06-02 loftytopping/PyBox#3 @loftytopping
Yes, sorry. Ive now added the v1.0.0 release against the master branch.
