### 2018-06-06 c29cd9c
Changed UManSysProp default link location to work with docker files
