+++ Aerosol/f2py/Fixed_yield/Aerosol_sim_fixed_yield.py
@@ -49,7 +49,7 @@ sys.path.append(os.path.abspath('../../..'))
 sys.path.append(os.path.abspath('../..'))
 import Parse_eqn_file # [•] Needed to parse the .eqn file, name given in this file
 import rate_coeff_conversion # [•] Converts standard text rate coefficients into Numba/Fortran
-import MCM_constants # [•] holds more pre-defined rate coefficients and photolysis rates not provided in .eqn file. 
+#import MCM_constants # [•] holds more pre-defined rate coefficients and photolysis rates not provided in .eqn file. 
 import Size_distributions # [•] Create size distribution according to number of bins and core-material
 import collections
 import pdb
