### 2018-06-07 18fe0d0
Now added an exception so that the suite runs in a docker container without raising an error if trying to plot
