+++ Aerosol/f2py/Fixed_yield/ODE_solver_fixedyield.py
@@ -291,6 +291,8 @@ def run_simulation(filename, save_output, start_time, temp, RH, H2O, PInit, y_co
     #Plot the change in concentration over time for a given specie. For the user to change / remove
     #In a future release I will add this as a seperate module
     if with_plots:
-        
-        Plotting.stacked_bar(t_array,y_matrix,num_species_condensed,num_bins,numpy.array(y_mw[include_index]),NA)
+        try:
+            Plotting.stacked_bar(t_array,y_matrix,num_species_condensed,num_bins,numpy.array(y_mw[include_index]),NA)
+        except:
+            print("There is a problem using Matplotlib in your environment. If using this within a docker container, you will need to transfer the data to the host or configure your container to enable graphical displays. More information can be found at http://wiki.ros.org/docker/Tutorials/GUI ")
         
