+++ f2py/ODE_solver.py
@@ -231,12 +231,16 @@ def run_simulation(filename, start_time, save_output, temp, RH, RO2_indices, H2O
     #Plot the change in concentration over time for a given specie. For the user to change / remove
     #In a future release I will add this as a seperate module
     if with_plots:
-        P.plot(t_array,numpy.log10(y_matrix[:,species_dict2array['APINENE']]), marker='o',label="APINENE")
-        P.plot(t_array,numpy.log10(y_matrix[:,species_dict2array['PINONIC']]), marker='o',label="PINONIC")
-        P.title(exp_mod.name)
-        P.legend(loc='upper left')
-        P.ylabel("Concetration log10[molecules/cc]")
-        P.xlabel("Time [seconds] since start of simulation")
-        P.show()
+
+        try:
+            P.plot(t_array,numpy.log10(y_matrix[:,species_dict2array['APINENE']]), marker='o',label="APINENE")
+            P.plot(t_array,numpy.log10(y_matrix[:,species_dict2array['PINONIC']]), marker='o',label="PINONIC")
+            P.title(exp_mod.name)
+            P.legend(loc='upper left')
+            P.ylabel("Concetration log10[molecules/cc]")
+            P.xlabel("Time [seconds] since start of simulation")
+            P.show()
+        except:
+            print("There is a problem using Matplotlib in your environment. If using this within a docker container, you will need to transfer the data to the host or configure your container to enable graphical displays. More information can be found at http://wiki.ros.org/docker/Tutorials/GUI ")
         
     
