### 2018-06-11 #755 @lheagy
@highendcompute informed me of a potential conflict of interest as he has published with @loftytopping in the past two years and has ongoing projects (that are not directly related to this one). As we have two reviewers, we will waive the conflict of interest and continue to have @highendcompute as a reviewer here. Please let me know if you have any concerns.
