### 2018-06-13 #755 @loftytopping
This sounds fine to me, many thanks!
I will be using PyBox with students in our faculty for their final year projects. Some are starting over the summer and I have created some additional instructions on how to run PyBox in a Docker container. Im guessing I should now wait until the review has finished before uploading these to the repository? Absolutely fine, of course, if so.
