### 2018-06-13 #755 @dvalters
@loftytopping Normally in the past for JOSS I have just reviewed the linked tagged release, and the master or other branches can be continued to be worked on as normal
