### 2018-06-18 3cccbdd
Added Dockerfile to run PyBox within a container based on an Ubuntu:16 image
