### 2018-06-18 35abd56
Added Dockerfile to run PyBox within a container based on an Ubuntu:16 image
