### 2018-06-19 #755 @loftytopping
Hi @lheagy Im on holiday from next Sunday and then back in work, following a conference, 11th July. Im just going through open papers and editorial jobs Im involved in to just let everyone know so there was no unsuspecting silence from me. Thnx, Dave.
