### 2018-07-11 fad9aa1
Taken out an error in the pre-written fortran module which confused compiler in Docker. Also removed the requirement for static compilation that was causing problems in Docker
