+++ f2py/f2py_rate_coefficient.py
@@ -35,7 +35,6 @@ ext = Extension (name = "rate_coeff_f2py",
                                      "-ffast-math",
                                      "-fopenmp"],
                  extra_link_args=["-lgomp",
-                                  "-static",
                                   "-llapack",
                                   "-lblas"])
 if __name__ == '__main__':
