+++ f2py/f2py_loss_gain.py
@@ -34,7 +34,6 @@ ext = Extension (name = "loss_gain_f2py",
                                      "-ffast-math",
                                      "-fopenmp"],
                  extra_link_args=["-lgomp",
-                                  "-static",
                                   "-llapack",
                                   "-lblas"])
 if __name__ == '__main__':
