### 2018-07-19 20be6d6
Changed default OpenMP flag to false on account of overhead analysis needed before general reccomendations
