+++ Parse_eqn_file.py
@@ -1144,27 +1144,79 @@ def write_rate_file_numba(filename,rate_dict):
     
     f.write('    KMT18 = 9.5E-39*O2*numba_exp(5270.0/temp)/(1+7.5E-29*O2*numba_exp(5610.0/temp))\n') 
 
+    #f.write('    # ************************************************************************\n') 
+    #f.write('    # define photolysis reaction rates using derwent method from mcm2box.fac\n') 
+    #f.write('    # ************************************************************************\n') 
+
+    #f.write('    # solar declination angle \n') 
+    #f.write('    dec = 23.79\n') 
+    #f.write('    # latitude\n') 
+    #f.write('    lat = 50.0\n') 
+    #f.write('    pi = 4.0*numba_arctan(1.0)\n') 
+    #f.write('    # local hour angle - representing time of day\n') 
+    #f.write('    lha = (1.0+ttime/4.32E+4)*pi\n') 
+    #f.write('    radian = 180.0/pi\n') 
+    #f.write('    lat = lat/radian\n') 
+    #f.write('    dec = dec/radian\n') 
+    #f.write('    theta = numba_arccos(numba_cos(lha)*numba_cos(dec)*numba_cos(lat)+numba_sin(dec)*numba_sin(lat))\n') 
+    #f.write('    sinld = numba_sin(lat)*numba_sin(dec)\n') 
+    #f.write('    cosld = numba_cos(lat)*numba_cos(dec)\n') 
+    #f.write('    cosx = (numba_cos(lha)*cosld)+sinld\n') 
+    #f.write('    cosx = numba_cos(theta)\n') 
+    #f.write('    secx = 1.0E+0/(cosx+1.0E-30)\n') 
+
     f.write('    # ************************************************************************\n') 
-    f.write('    # define photolysis reaction rates using derwent method from mcm2box.fac\n') 
+    f.write('    # define photolysis reaction rates using the AtChem2 model which can be found: \n') 
+    f.write('    # https://github.com/AtChem/AtChem2 \n')
     f.write('    # ************************************************************************\n') 
 
-    f.write('    # solar declination angle \n') 
-    f.write('    dec = 23.79\n') 
-    f.write('    # latitude\n') 
-    f.write('    lat = 50.0\n') 
-    f.write('    pi = 4.0*numba_arctan(1.0)\n') 
-    f.write('    # local hour angle - representing time of day\n') 
-    f.write('    lha = (1.0+ttime/4.32E+4)*pi\n') 
-    f.write('    radian = 180.0/pi\n') 
-    f.write('    lat = lat/radian\n') 
-    f.write('    dec = dec/radian\n') 
-    f.write('    theta = numba_arccos(numba_cos(lha)*numba_cos(dec)*numba_cos(lat)+numba_sin(dec)*numba_sin(lat))\n') 
-    f.write('    sinld = numba_sin(lat)*numba_sin(dec)\n') 
-    f.write('    cosld = numba_cos(lat)*numba_cos(dec)\n') 
-    f.write('    cosx = (numba_cos(lha)*cosld)+sinld\n') 
-    f.write('    cosx = numba_cos(theta)\n') 
-    f.write('    secx = 1.0E+0/(cosx+1.0E-30)\n') 
+    f.write('    pi = 4.0*numba_arctan(1.0) \n') 
 
+    f.write('    currentDayOfYear=150.0 \n') 
+    f.write('    latitude=51.51  \n') 
+    f.write('    longitude=0.13  \n') 
+    f.write('    cosx_threshold = 1.0E-30  \n') 
+
[…]
