### 2018-07-31 #755 @loftytopping
Hi guys. Is it possible to have a projected update? I have a couple of groups starting to use it and approach study draft publications. Thanks!
