### 2018-07-31 #755 @dvalters
Will resume the last bits of testing tonight - I've ticked off the checklist above where I believe it passes so far (mostly really good - don't see any major barriers to it being accepted).

My only comment really at the minute is on the installation procedure. Outwith the docker image, it's fairly tricky to install. Just been trying to install it on a university Linux machine this morning and without docker/conda it's proved tricky - had issues getting Assimulo to work, linking to the right libs etc..

Having a `setup.py` file or as a minimum a `requirements.txt` file might help some of these issues? (For example, the dependencies section of the documentation lists third party modules as well as built-in modules (like `shutil` and `multiprocessing`, and also doesn't specify whether the third party modules have minimum versions.) A conda installation would be great although I realise that's probably quite a bit of extra work.
