### 2018-07-31 loftytopping/PyBox#4 @dvalters
Running test_modules.py script results in error?
Running the `test_modules.py` module results in the following output:

```
(py36) [dav@ncc1701d test]$ python test_modules.py -v
Opening file  MCM_APINENE.eqn.txt for parsing
Calculating total number of equations =  836
Parsing each equation
Total number of species =  305
Saving all equation information to dictionaries
Mapping species names to SMILES and Pybel objects
No SMILES entry for species  NO3
No SMILES entry for species  CO
No SMILES entry for species  H2
No SMILES entry for species  HNO3
No SMILES entry for species  NO
No SMILES entry for species  NO2
No SMILES entry for species  SO2
No SMILES entry for species  SO3
No SMILES entry for species  H2O2
Traceback (most recent call last):
  File "test_modules.py", line 446, in <module>
    setup(filename)
  File "test_modules.py", line 111, in setup
    reaction_dict=outputdict['reaction_dict']
KeyError: 'reaction_dict'
```

Ref: https://github.com/openjournals/joss-reviews/issues/755
