### 2018-08-01 b132568
Corrected problem with testing suite in response to JOSS review. Trying to extract dictionary that dosnt exist anymore and isnt required for testing. Also noticed the testing suite was using data from a previous setup and thus slightly different. This has been changed.
