### 2018-08-01 da26406
Changed the initial concentrations back to the original conditions
