### 2018-08-01 3bcebf6
Made changes in response to JOSS review. In particular sorted out assert equal issues between numba and fotran files
