### 2018-08-02 loftytopping/PyBox#4 @loftytopping
Many thanks for ponting this out. Id left a component dictionary in the script that is used for testing purposes. This has now been removed.
