### 2018-08-02 #755 @loftytopping
Many thanks @dvalters for spending time on this. As for the general installation procedure, the problem is for sure with Assimulo. It is also platform depdenent. Im not sure I could provide a generic script to ensure this is installed automatically as, as you say, there is a reliance on library locations etc. I completely agree that this could be better discussed and, if its ok with you, I will do this now as a solution to any perceived problem.
