### 2018-08-02 #755 @dvalters
@loftytopping Sure that's fine. It's not a JOSS requirement/barrier to acceptance to have an automated setup and as long as the instructions are detailed enough that is fine in my opinion.
