### 2018-08-02 ec654d2
Update README.md
Changed in response to JOSS review
