### 2018-08-02 #755 @highendcompute
There are better options (generally I speak) than a bash script for platform independent installs based on dependencies, but it seems this isn’t a necessary requirement. The likes of

I have built using the Docker provided – which worked when I had sufficient space (so maybe need to give heads up on requirements?) and which was larger than expected but I don’t think it’s a requirement to provide the most stream line docker (eg how one does the layers matters; a trimmed down Linux to start with; and other ideas). I just need more time so can focus and complete my review. M
