+++ docs/index.md
@@ -9,9 +9,11 @@ This project is licensed under the terms of the GNU General Public License v3.0,
 # Table of contents
 1. [Model overview](#Model-overview)
 2. [Dependencies and installation](#Dependencies)
-     2a. [Using a Docker container](#Docker)
+     * 2a. [Conda](#Conda)
+     * 2b. [Using a Docker container](#Docker)
+     * 2c. [Alternative](#Alternative)
 3. [Folder structure and running the model](#Folder-Structure)
-4. [Unit Tests](#Automated-unit-tests)
+4. [Unit tests](#Automated-unit-tests)
 5. [Contributing](#Contributing)
 6. [Code of Conduct](#Code-of-Conduct)
 
@@ -32,23 +34,97 @@ Where the equation number is defined first, then the reactants/products along wi
 
 ## Dependencies and installation <a name="Dependencies"></a>
 
-PyBox has been built in the [Anaconda Python 3.6 environment](https://www.anaconda.com/download/#macos). Utilised modules included within the Anaconda environment are:
-
- - Numpy
- - Scipy
- - Matplotlib
- - unittest
- - shutil
- - xml.dom
- - multiprocessing
+PyBox has been built in the [Anaconda Python 3.6 environment](https://www.anaconda.com/download/#macos). [Assimulo](http://www.jmodelica.org/assimulo) is the numerical core of PyBox. The Assimulo Ordinary Differential Equation (ODE) solver package allows us to use solvers designed for stiff systems. [UManSysProp](http://umansysprop.seaes.manchester.ac.uk) is used to automate predictions of pure component and mixture properties to allow gas-to-particle partitioning simulations. Before methods of installation are described, below is a list of both 3rd party and Python native modules required. You can click on each to visit the project website with further details, including useful information on installation pertinent to the section on using an [alternative](#Alternative) installation procedure to either Docker or Conda. The version number reflects that which has been used/tested against the provided source code here. I also note those modules that are require specifically for the UManSysProp package noted earlier:
  
-Additional dependencies are:
+#### 3rd party modules:
 
-- [Assimulo](http://www.jmodelica.org/assimulo). This is the numerical core of PyBox. The Assimulo Ordinary Differential Equation (ODE) solver package allows us to use solvers designed for stiff systems. As found on the project website, there are multiple [methods for installation](https://jmodelica.org/assimulo/installation.html), including both package managers and compiling from source.  From my own experience, it is better to build from source against your Anaconda Python environment. You will need to point to the location of the [Sundials solver suite](https://computation.llnl.gov/projects/sundials) and both BLAS and LAPACK. You can check if your Assimulo installation has worked by opening an interactive Python shell and typing:
+Module        |  Version [dependency]
+------------- | -------------
+[Assimulo](https://jmodelica.org/assimulo/)      |  2.9
+[flask-wtf](http://flask-wtf.readthedocs.io/en/stable/)     |  0.14.2 [UManSysProp]
+[matplotlib](https://matplotlib.org/)    |  2.0.2
+[numba](https://numba.pydata.org/)         |  0.33.0 
+[numpy](https://www.scipy.org/scipylib/download.html)         |  1.12.1
+[openbabel python interface](https://openbabel.org/docs/dev/UseTheLibrary/Python_Pybel.html)     |  2.4.1  [UManSysProp] 
+[pandas](https://pandas.pydata.org/)        |  0.20.1
+[scipy](https://www.scipy.org/scipylib/download.html)         |  0.19.1
+[wtforms](https://wtforms.readthedocs.io/en/stable/)       |  2.1    [UManSysProp]
+
+#### Built-in modules
+
[…]
