### 2018-08-02 #755 @loftytopping
Sure, thanks! @dvalters ive now changed the README.md file to include more information on the dependencies according to Conda/Docker/other. Hopefully this adds more clarity to any installation procedure. @lheagy Im making changes to the current master branch, will any potential paper point to a new version specific once everything has been approved? thanks!
