### 2018-08-03 loftytopping/PyBox#5 @dvalters
Minor comments on the readme file
Sorry...these are really tiny comments:

1. Just missing a `.py` on the end here:

> Aerosol
> In the Aerosol folder you can find gas-to-particle partitioning frameworks. There are two examples provided. The first, within the f2py folder, simulates the partitioning of compounds to 16 size bins again from the Alpha-Pinene chemical mechanism as this evolves over time. This uses properties calculated from the UManSysProp suite. To run the model, once you are happy all dependencies are installed, type the following from the Aerosol/f2py directory:
>
> python Aerosol_simulation_f2py**.py**    <--- Missing a .py extension
>
> In addition to the species concentrations and ambient conditions, you can change the size distribution and number of size bins in the following:

2. I would add `f2py` as a dependency as not everyone might be using Anaconda (I use Miniconda for example and you don't get very many, if any, third party packages outside core Python)

Ref: openjournals/joss-reviews/issues/755
