### 2018-08-03 loftytopping/PyBox#4 @loftytopping
Ah, yes. Given the huge range of component concetrations, I will need to re-factor the error checking. Apologies. I will try to get round to checking whether this is appropriate today before mid-afternoon. Thanks.
