### 2018-08-03 loftytopping/PyBox#5 @loftytopping
Sure, no problem. I will make these changes.
