### 2018-08-03 3e8da43
Changed the testing method to use Numpy's allclose with a relative tolerance
