### 2018-08-03 6b61ee6
Update README.md
Made changes in response to JOSS review
