### 2018-08-03 loftytopping/PyBox#4 @loftytopping
Thanks for noticing this. I should have been using the 'numpy.allclose' function with a defined relative error. I didnt realise that the using 'numpy.testing.assert_almost_equal' would create errors based on defining only absolute errors. This will be a general problem for any model with a wide range of numeric values. I have made the change to using numpy.allclose.
