### 2018-08-03 #755 @dvalters
@loftytopping Just seen the latest fixes - thanks - will have a final lookover this evening, looks good to go otherwise
