### 2018-08-03 #755 @highendcompute
I pulled updates as of 3rd Aug, stopped and removed all previous pybox containers, removed all previous pybox, <none> & ubuntu images (presume <none> was intermediate later). From this point:

for the Dockerfile I had to add an extra "no" to prevent hanging on final question:
```
Do you wish the installer to prepend the Anaconda3 install location
to PATH in your /root/.bashrc ? [yes|no]
[no] >>>
Appending source /root/anaconda3/bin/activate to /root/.bashrc
A backup will be made to: /root/.bashrc-anaconda3.bak

For this change to become active, you have to open a new terminal.

Thank you for installing Anaconda3!

===========================================================================

Anaconda is partnered with Microsoft! Microsoft VSCode is a streamlined
code editor with support for development operations like debugging, task
running and version control.

To install Visual Studio Code, you will need:
  - Administrator Privileges
  - Internet connectivity

Visual Studio Code License: https://code.visualstudio.com/license

Do you wish to proceed with the installation of Microsoft VSCode? [yes|no]
>>>
```

my general comments in terms of optimising a user's build via docker:
1) is it possible to use miniconda then build up only what is required? (base ubuntu vmem 83 MB to final layer being 4.4 GB)
2) It is recommended for Docker containers to NOT run as root, but as a user (set up in Dockerfile)

please note that I am happy with this current version meets the JOSS requirements, the above are merely suggestions/questions for potential improvements in future versions

I will open other comments re running the Docker image and on building outside of Docker
