### 2018-08-03 #755 @lheagy
Hi @loftytopping, to answer your question - any changes made to the paper before we accept it will be included in the published version
