### 2018-08-04 #755 @dvalters
> "conda install openbabel" however fails to find the required package.

@highendcompute I had to do `conda install -c openbabel openbabel` (i.e. it's not in the main conda repository)
