### 2018-08-06 65be3ef
Removed character in Numba Jacobian file causing Docker to fail
