### 2018-08-06 #755 @loftytopping
@highendcompute thanks for the note re the Dockerfile. Ive now corrected this, it seems that an extra character prior to an EOL statement in the Numba file crept in when editing on my Mac. This has now been removed and tested in a new container. Ive now changed the README file and docs Index.md to include the full openbabel repository. More on the other items ASAP.
