### 2018-08-06 #755 @loftytopping
The addition of an extra "no" in the printf chain is not required and raises an error message on build completion. Unfortunately the MS Visual Studio component section is simply too slow, and I'm not sure why. It appears to hang, but does proceed if left. I will add a note on this in the Dockerfile readme attachment.
