### 2018-08-07 #755 @highendcompute
regarding the Dockerfile and my orginal comment about needing additional "no" in the printf chain. I have investigated and it appears this is not an issue for the Dockerfile but rather running Docker under a Linux VM with limited disk & memory attributes leads to extortionate long waits (typically at the point of removing intermediate containers). It is not necessary for this "paper" but future considerations may be to reduce the number of such and to use the 'batch' mode for Anaconda, eg
RUN wget https://repo.continuum.io/archive/Anaconda3-5.1.0-Linux-x86_64.sh && bash -x Anaconda3-5.1.0-Linux-x86_64.sh -b && rm Anaconda3-5.1.0-Linux-x86_64.sh
