### 2018-08-07 #755 @highendcompute
i can now run the docker version.
i will check the handbuild version tonight
