### 2018-08-07 #755 @loftytopping
Thanks @highendcompute . In response to your concerns, Ive added the following information,licensing and note on future builds to the Dockerfile. Based on Docker forums it seems this root issue is an ongoing debate:

#                                                                                        #
##########################################################################################
# procedures                                                                             #
##########################################################################################
