### 2018-08-07 #755 @loftytopping
Ive also now created an enhancement request on the project repo as : Lightweight Docker image #6
Ive assigned myself to sort this over the near future.
