### 2018-08-07 1076910
Added more information on Docker build in response to JOSS review
