### 2018-08-09 #755 @highendcompute
Hi, not sure why all these different comments/suggestions are in a single thread (@lheagy ?) but here's my penultimate comments:
* docker: builds and runs as expected. (I cannot try the OpenMP version of this due to local set-up)
* non-docker: for my DEBIAN set-up, I chose the option for anaconda to handle dependencies (probably should note this is incompatible with the usual DEBIAN package management set-up), specifically for python3. this builds and runs almost as expected: one nit-pick point that the numba needs re-building each time and the Gas_simulation.py does not seem to terminate once the plot is created. Nice touch having the plot.
* when I did attempt to build via DEBIAN dependencies, and Python 2 (sic) I hit problems with special characters in the Gas_simulation.py so suggest recommendation this code is for Python 3 only

On the parallel front, it is very good to see this included. I cannot test on my docker build since that is within a VM and the infrastructure does not support good parallel performance (tests done elswehere)

I shall now look at a CLI build of PyBox on native Linux system on multicore machine in order to feedback.

M
