### 2018-08-09 #755 @highendcompute
NB on some systems, going anaconda route I got some issues, as highlighted in the README.md as likely need install sundials explicitly, which I did via
 conda install -c chemreac sundials blas
but still get

 [15:48:51] ~/tmp/PyBox$ python
Python 3.6.4 |Anaconda, Inc.| (default, Jan 16 2018, 18:10:19)
[GCC 7.2.0] on linux
>>> from assimulo.solvers import RodasODE, CVode
Could not find  not found (required by /mnt/iusers01/pp01/mccssmb2/anaconda3/lib/python3.6/site-packages/assimulo/lib/radau5.cpython-36m-x86_64-linux-gnu.so)
Could not find  not found (required by /mnt/iusers01/pp01/mccssmb2/anaconda3/lib/python3.6/site-packages/assimulo/solvers/sundials.cpython-36m-x86_64-linux-gnu.so)
Could not find  not found (required by /mnt/iusers01/pp01/mccssmb2/anaconda3/lib/python3.6/site-packages/assimulo/solvers/kinsol.cpython-36m-x86_64-linux-gnu.so)
Could not find  not found (required by /mnt/iusers01/pp01/mccssmb2/anaconda3/lib/python3.6/site-packages/assimulo/lib/dopri5.cpython-36m-x86_64-linux-gnu.so)
Could not find  not found (required by /mnt/iusers01/pp01/mccssmb2/anaconda3/lib/python3.6/site-packages/assimulo/lib/rodas.cpython-36m-x86_64-linux-gnu.so)
Could not find  not found (required by /mnt/iusers01/pp01/mccssmb2/anaconda3/lib/python3.6/site-packages/assimulo/lib/odassl.cpython-36m-x86_64-linux-gnu.so)
Could not find ODEPACK functions
Could not find RADAR5
Could not find GLIMDA
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
ImportError: cannot import name 'RodasODE'

(This has happened on two systems where there was a system anaconda but it wasn't possible to use for obtaining new packages (since tried to install in system dir) so used the Anaconda script directly to install in $HOME/anaconda3 and set PATH to include this as first element
