### 2018-08-09 #755 @lheagy
@highendcompute: These are all in one thread as you are replying on the JOSS review. I will transfer your comments to issues on the target repository, https://github.com/loftytopping/PyBox, so that @loftytopping can respond to them individually there.
