### 2018-08-09 loftytopping/PyBox#7 @lheagy
continued...

running from docker image, i appear to hit a problem:

`
 [20:48:10] ~/DT_pybox/PyBox$ docker run  -it pybox
root@01a532291b93:/Code# whoami
root
root@01a532291b93:/Code#  cd /Code/Git_repos/PyBox/
root@01a532291b93:/Code/Git_repos/PyBox# python Gas_simulation.py
Opening file  MCM_APINENE.eqn.txt for parsing
Calculating total number of equations =  836
Parsing each equation
Total number of species =  305
Saving all equation information to dictionaries
Saving the mechanism dictionary as a pickled object for later retrieval
Converting rate coefficient operations into Python-Numba file
Creating Python-Numba functions and sparse matrices for product multiplications and dydt function
Creating Python-Numba Jacobian
Traceback (most recent call last):
  File "Gas_simulation.py", line 163, in <module>
    Parse_eqn_file.write_gas_jacobian_numba(filename,equations,num_species,loss_dict,gain_dict,species_dict2array,rate_dict_reactants)
  File "/Code/Git_repos/PyBox/Parse_eqn_file.py", line 2987, in write_gas_jacobian_numba
    f.write('#\xa7# \n')
UnicodeEncodeError: 'ascii' codec can't encode character '\xa7' in position 1: ordinal not in range(128)
root@01a532291b93:/Code/Git_repos/PyBox#
`
