### 2018-08-09 loftytopping/PyBox#7 @lheagy
building on linux server (uname: Linux mkb-debian-vm 3.16.0-4-amd64 #1 SMP Debian 3.16.43-2+deb8u3 (2017-08-15) x86_64 GNU/Linux)

general comments (do not need fixing but please consider):
*) maybe state if python2 is okay or whether it has to be python3 (I'm testing with python2)
*) why both numpy and scipy? isn't it possible to use just one or the other

In terms of the dependencies for Python (not my primarily language I admit), Debian does not have any form of conda in its package manager suite nor do I see URL/.deb for alternative. It would be useful to clarify whether any Python package manager (eg pip) will suffice or whether there is a good reason to use anaconda.

In order to validate your assertions I have used a VM and run the anaconda install script. The following is based on this:
* with the one exception (next bullet) everything builds as stated in the README.md
* "conda install openbabel" actually requires "-c openbabel" as Declan V noted

In terms of testing this build, I again get a (de)coding error:

 [14:34:53] ~/DT_pybox/PyBox$ python Gas_simulation.py
  File "Gas_simulation.py", line 38
SyntaxError: Non-ASCII character '\xe2' in file Gas_simulation.py on line 38, but no encoding declared; see http://python.org/dev/peps/pep-0263/ for details
