### 2018-08-09 loftytopping/PyBox#7 @lheagy
@loftytopping, @highendcompute, from your conversation these issues look to be resolved? and superseded by #6. Please feel free to close this @loftytopping if that is the case
