### 2018-08-09 #755 @lheagy
@loftytopping, @highendcompute: I have transferred your comments to issues
- [ ]  loftytopping/PyBox#7
- [ ]  loftytopping/PyBox#8
Please feel free to add further comments and discussion there. From your conversation, it looks like loftytopping/PyBox#7 can be closed, but I will let you confirm.
