### 2018-08-12 7ad544c
Removed default option to use OpenMP and will add cautionary note to README
