### 2018-08-12 loftytopping/PyBox#7 @loftytopping
Yes thats right @lheagy many thanks. Ive now created a new enhancement issue for future development, to provide a 'lightweight' docker image.
