### 2018-08-12 loftytopping/PyBox#8 @loftytopping
Many thanks for your time and effort on this @highendcompute. This is unfortunately a sympton of the Assimulo package. In response, Ive made the following changes:
- Urged caution in the README instructions, even with the conda build.
- Created an enhancement issue to not just rely on Assimulo in PyBox. After rigirous profiling and model testing re the ODE solvers, and especially after new developments, I will be able to make the decision on whether I can rely on Scipy libraries which would remove these issues. The Scipy libraries currently have the CVode solvers, and may include other stiff solvers in a future release. With regards to this, I will add a request that the Rodas methods are implemented.

In addition, based on the above ethos, I have removed the default option to use OpenMP but leave this for a future version that has been fully profiled and optmised. This is also retained as an enhancement request and is duly noted in both the Python files and README.
