### 2018-08-13 #755 @highendcompute
with the resolution of all issues i am now happy with this to be published.
do i need do anything else at this stage? m
