### 2018-08-13 #755 @lheagy
Excellent, thanks @highendcompute! There is nothing else you need to do at this stage. Thank you for your work on this review!

@dvalters: there is one remaining item on your checklist: **functionality**. Are there outstanding issues to be addressed here? Thanks!
